Rearranged folders for exceptions and handlers Rearranged folders for test classes
@Transactional annotations and try/catch blocks in services fixed
